Move some tests to test_algos, test_hyper_map, test_raise_assert
